Move importers to a separate Django app
