+++ importers/tests/test_data/archlinux.json
@@ -1,0 +1,34 @@
+[
+  {
+    "name": "AVG-708",
+    "packages": [
+      "wireshark-common",
+      "wireshark-gtk",
+      "wireshark-cli",
+      "wireshark-qt"
+    ],
+    "status": "Fixed",
+    "severity": "Critical",
+    "type": "multiple issues",
+    "affected": "2.6.0-1",
+    "fixed": "2.6.1-1",
+    "ticket": null,
+    "issues": [
+      "CVE-2018-11362",
+      "CVE-2018-11361",
+      "CVE-2018-11360",
+      "CVE-2018-11359",
+      "CVE-2018-11358",
+      "CVE-2018-11357",
+      "CVE-2018-11356",
+      "CVE-2018-11355",
+      "CVE-2018-11354"
+    ],
+    "advisories": [
+      "ASA-201805-25",
+      "ASA-201805-24",
+      "ASA-201805-22",
+      "ASA-201805-23"
+    ]
+  }
+]
