+++ importers/tests/test_data/debian.json
@@ -1,0 +1,150 @@
+{
+  "librsync": {
+    "CVE-2014-8242": {
+      "scope": "remote",
+      "debianbug": 776246,
+      "description": "librsync before 1.0.0 uses a truncated MD4 checksum to match blocks",
+      "releases": {
+        "stretch": {
+          "status": "open",
+          "nodsa_reason": "",
+          "repositories": {
+            "stretch": "0.9.7-10"
+          },
+          "nodsa": "Minor issue, too instrusive to backport",
+          "urgency": "low"
+        },
+        "jessie": {
+          "status": "open",
+          "nodsa_reason": "",
+          "repositories": {
+            "jessie": "0.9.7-10"
+          },
+          "nodsa": "Minor issue, too instrusive to backport",
+          "urgency": "low"
+        },
+        "sid": {
+          "status": "resolved",
+          "repositories": {
+            "sid": "2.0.2-1"
+          },
+          "fixed_version": "2.0.2-1",
+          "urgency": "low"
+        },
+        "bullseye": {
+          "status": "resolved",
+          "repositories": {
+            "bullseye": "2.0.2-1"
+          },
+          "fixed_version": "2.0.2-1",
+          "urgency": "low"
+        },
+        "buster": {
+          "status": "open",
+          "nodsa_reason": "",
+          "repositories": {
+            "buster": "0.9.7-10"
+          },
+          "nodsa": "Minor issue, too instrusive to backport",
+          "urgency": "low"
+        }
+      }
+    }
+  },
+  "mimetex": {
+    "CVE-2009-1382": {
+      "scope": "remote",
+      "debianbug": 537254,
+      "description": "Multiple stack-based buffer overflows in mimetex.cgi in mimeTeX",
+      "releases": {
+        "stretch": {
+          "status": "resolved",
+          "repositories": {
+            "stretch": "1.74-1"
+          },
+          "fixed_version": "1.50-1.1",
+          "urgency": "medium"
+        },
+        "jessie": {
+          "status": "resolved",
+          "repositories": {
+            "jessie": "1.74-1"
+          },
+          "fixed_version": "1.50-1.1",
+          "urgency": "medium"
+        },
+        "sid": {
+          "status": "resolved",
+          "repositories": {
+            "sid": "1.76-1"
+          },
+          "fixed_version": "1.50-1.1",
+          "urgency": "medium"
+        },
+        "bullseye": {
+          "status": "resolved",
+          "repositories": {
+            "bullseye": "1.76-1"
+          },
+          "fixed_version": "1.50-1.1",
+          "urgency": "medium"
+        },
+        "buster": {
+          "status": "resolved",
+          "repositories": {
+            "buster": "1.76-1"
+          },
+          "fixed_version": "1.50-1.1",
+          "urgency": "medium"
+        }
+      }
+    },
+    "CVE-2009-2459": {
+      "scope": "remote",
+      "debianbug": 537254,
+      "description": "Multiple unspecified vulnerabilities in mimeTeX",
+      "releases": {
+        "stretch": {
+          "status": "resolved",
+          "repositories": {
+            "stretch": "1.74-1"
+          },
+          "fixed_version": "1.50-1.1",
+          "urgency": "medium"
+        },
+        "jessie": {
+          "status": "resolved",
+          "repositories": {
+            "jessie": "1.74-1"
+          },
+          "fixed_version": "1.50-1.1",
+          "urgency": "medium"
+        },
+        "sid": {
+          "status": "resolved",
+          "repositories": {
+            "sid": "1.76-1"
+          },
+          "fixed_version": "1.50-1.1",
+          "urgency": "medium"
+        },
+        "bullseye": {
+          "status": "resolved",
+          "repositories": {
+            "bullseye": "1.76-1"
+          },
+          "fixed_version": "1.50-1.1",
+          "urgency": "medium"
+        },
+        "buster": {
+          "status": "resolved",
+          "repositories": {
+            "buster": "1.76-1"
+          },
+          "fixed_version": "1.50-1.1",
+          "urgency": "medium"
+        }
+      }
+    }
+  }
+}
